ViaMichelin Key erneuert und Distance bei Route mit reingenommen

+++ src/main/java/travelcompare/restapi/external/viamichelin/model/RoadSheetStep.java
@@ -24,7 +24,7 @@ public class RoadSheetStep {
 
     @Getter
     @Setter
-    private Double distance;
+    private Long distance;
 
     @Getter
     @Setter
